requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
boto3==1.24.41
botocore==1.27.96
cffi==1.16.0
cycler==0.12.1
defusedxml==0.7.1
diff-match-patch==20230430
Django==3.2.15
django-ckeditor==6.3.2
django-cors-headers==3.13.0
django-environ==0.9.0
django-import-export==2.8.0
django-js-asset==2.2.0
django-storages==1.12.3
djangorestframework==3.13.1
et-xmlfile==1.1.0
fonttools==4.53.1
gunicorn==20.1.0
jmespath==1.0.1
joblib==1.4.2
kiwisolver==1.4.5
matplotlib==3.5.3
numpy==1.23.2
odfpy==1.4.1
openpyxl==3.1.5
packaging==24.1
pandas==1.4.3
Pillow==9.2.0
psycopg2==2.9.3
pycparser==2.22
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
s3transfer==0.6.2
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
sqlparse==0.4.4
tablib==3.6.1
threadpoolctl==3.5.0
typing_extensions==4.10.0
urllib3==1.26.18
whitenoise==6.2.0
xlrd==2.0.1
xlwt==1.3.0
